tests pass, course part 4 complete

+++ controllers/blogs.js
@@ -1,6 +1,8 @@
 const blogsRouter = require('express').Router()
+const jwt = require('jsonwebtoken')
 const Blog = require('../models/blog')
 const User = require('../models/user')
+
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog
@@ -8,11 +10,16 @@ blogsRouter.get('/', async (request, response) => {
   response.json(blogs)
 })
 
+
 blogsRouter.post('/', async (request, response) => {
   const body = request.body
 
-  const user = await User.findOne({})
-
+  const token = request.token
+  const decodedToken = jwt.verify(token, process.env.SECRET)
+  if (!token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
   const blog = new Blog({
     title: body.title,
     author: body.author,
@@ -30,8 +37,30 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  //TODO refactor so this code isn't redundant
+  const token = request.token
+  const blog = await Blog.findById(request.params.id)
+  const decodedToken = jwt.verify(token, process.env.SECRET)
+
+  if (!token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+  const userid = await (await User.findById(decodedToken.id))._id
+
+  if (!blog) {
+    return response.status(401).json({ error: 'blog not found' })
+  } else if (!userid || !blog.user) {
+    return response.status(401).json({ error: 'you do not have permission to delete this blog' })
+  }
+
+  if (blog.user.toString() === userid.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  } else {
+    return response.status(401).json({ error: 'you do not have permission to delete this blog' })
+  }
+
+
 
 })
